Create new module & reorganize helper methods
