+++ lib/game.rb
