+++ lib/gameteam.rb
