+++ lib/goals_utility.rb
@@ -1,16 +1,8 @@
-module Seasonable
-
-  def all_games_by_season
-    @all_games_by_season ||= @games.group_by { |game| game.season } 
-  end
+module GoalsUtility
 
   def sums_of_home_away_goals
     @sums_of_home_away_goals ||= @games.map { |game| (game.away_goals + game.home_goals)}
   end
-
-  # def team_id_all_goals
-  #   @team_id_all_goals ||= @game_teams.group_by { |game_teams| [game_teams.team_id] << game_teams.goals.to_f }
-  # end
 
   def team_id_all_goals
     team_id_all_goals_hash = Hash.new { |hash, key| hash[key] = [] }
@@ -20,9 +12,4 @@ module Seasonable
     @team_id_all_goals ||= team_id_all_goals_hash
   end
 
-  def team_name(id)
-    @teams.each do |team|
-        return team.team_name if team.team_id == id 
-    end
-  end
 end
