+++ spec/spec_helper.rb
@@ -2,13 +2,18 @@ require 'simplecov'
 SimpleCov.start
 
 require './lib/stat_tracker'
+
 require './lib/game_stats'
 require './lib/league_stats'
 require './lib/season_stats'
 require './lib/team_stats'
-require './lib/obj_game'
-require './lib/obj_gameteam'
-require './lib/obj_team'
-require './lib/statistics_module'
+
 require './lib/stats'
-require './lib/seasonable'
+
+require './lib/game'
+require './lib/gameteam'
+require './lib/team'
+
+require './lib/statistics_module'
+require './lib/season_utility'
+require './lib/goals_utility'
